+++ vite.config.ts
@@ -1,3 +1,4 @@
+/// <reference types="vitest/config" />
 import tailwindcss from "@tailwindcss/vite";
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
@@ -7,5 +8,9 @@ export default defineConfig({
 	server: {
 		port: 3000,
 		allowedHosts: ['ignignokt', 'ignignokt.home.youngsecurity.net']
+	},
+	test: {
+		include: ['src/**/*.test.ts'],
+		environment: 'node'
 	}
 });